End of Section 5

+++ Src/AuctionService/Controllers/AuctionsController.cs
@@ -54,7 +54,7 @@ namespace AuctionService.Controllers
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -78,7 +78,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            //if (auction.Seller != User.Identity.Name) return Forbid();
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -94,7 +94,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
-        //[Authorize]
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -102,7 +102,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            //if (auction.Seller != User.Identity.Name) return Forbid();
+            if (auction.Seller != User.Identity.Name) return Forbid();
             _context.Auctions.Remove(auction);
             await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
